+++ starboard/android/shared/video_decoder.h
@@ -163,7 +163,7 @@ class MediaCodecVideoDecoder : public VideoDecoder,
       decode_target_graphics_context_provider_;
   const std::string max_video_capabilities_;
   const std::optional<int> initial_max_frames_in_decoder_;
-  const int max_pending_inputs_size_;
+  const size_t max_pending_inputs_size_;
 
   // Android doesn't officially support multi concurrent codecs. But the device
   // usually has at least one hardware decoder and Google's software decoders.
